Organize files and remove unnecessary classes
