--- OpenTabletDriver/Models/SerializableDictionary.cs
@@ -1,37 +1,0 @@
-using System.Collections.Generic;
-using System.Xml;
-using System.Xml.Schema;
-using System.Xml.Serialization;
-
-namespace OpenTabletDriver.Models
-{
-    [XmlRoot("SerializableDictionary")]
-    public class SerializableDictionary<TKey, TValue> : Dictionary<TKey, TValue>, IXmlSerializable
-    {
-        public XmlSchema GetSchema() => null;
-
-        public void ReadXml(XmlReader reader)
-        {
-            if(reader.IsEmptyElement) { return; }
-            reader.Read();
-            while (reader.NodeType != XmlNodeType.EndElement)
-            {
-                object key = reader.GetAttribute("Key");
-                object value = reader.GetAttribute("Value");
-                this.Add((TKey)key, (TValue)value);
-                reader.Read();
-            }
-        }
-
-        public void WriteXml(XmlWriter writer)
-        {
-            foreach (var key in this.Keys)
-            {
-                writer.WriteStartElement("KeyValuePair");
-                writer.WriteAttributeString("Key", key.ToString());
-                writer.WriteAttributeString("Value", this[key].ToString());
-                writer.WriteEndElement();
-            }
-        }
-    }
-}
